Team_Workspace/Cha_SH/Cha_260408/Main_Docker_Runtime/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
langchain
langchain-core
langchain-openai
langchain-google-genai
openai
python-dotenv
python-multipart
psycopg2-binary
GeoAlchemy2
shapely
unstructured
PyMuPDF
aiohttp

redis[hiredis]
PyJWT
passlib[bcrypt]
